requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.9.0
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
